Move util tests to the right package
